# apps/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Container Manager API
  description: |
    A Container Manager system that provides WebContainer-compatible APIs for Docker container management.
    The system creates ephemeral Docker containers that can be controlled via REST APIs and WebSocket connections,
    allowing seamless migration from WebContainers to Docker-based containers.

    ## Architecture
    - **Backend**: NestJS service managing Docker containers via Dockerode
    - **WebSocket**: Real-time container interactions with persistent directory state
    - **Dual API**: Direct container management + WebContainer-compatible endpoints
  version: "1.0.0"
  contact:
    name: Container Manager API
    url: http://localhost:3001/api
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Development server
...
paths:
...
  /containers:
    get:
      tags:
        - containers
      summary: Get all containers
      description: Retrieve a list of all containers in the system
      operationId: getAllContainers
      responses:
        '200':
          description: All containers retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContainerInfo'
    post:
      tags:
        - containers
      summary: Create a new container
      description: Create a new Docker container with optional initial files
      operationId: createContainer
      requestBody:
        description: Container creation options
        content:
          application/json:
            schema:
              type: object
              properties:
                files:
                  $ref: '#/components/schemas/FileSystemTree'
              example:
                files:
                  "package.json":
                    file:
                      contents: |
                        {
                          "name": "my-project",
                          "version": "1.0.0"
                        }
                  "src":
                    directory:
                      "index.js":
                        file:
                          contents: "console.log('Hello World');"
      responses:
        '201':
          description: Container created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerInfo'
        '500':
          description: Failed to create container
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /containers/{id}:
    get:
      tags:
        - containers
      summary: Get container information
      description: Retrieve information about a specific container
      operationId: getContainer
      parameters:
        - name: id
          in: path
          required: true
          description: Container ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Container information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerInfo'
        '404':
          description: Container not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - containers
      summary: Delete a container
      description: Stop and remove a Docker container
      operationId: deleteContainer
      parameters:
        - name: id
          in: path
          required: true
          description: Container ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Container deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Container deleted successfully"
        '500':
          description: Failed to delete container
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /containers/{id}/url:
    get:
      tags:
        - containers
      summary: Get the preview URL for the container
      description: Retrieve the preview URL and port information for a container
      operationId: getContainerUrl
      parameters:
        - name: id
          in: path
          required: true
          description: Container ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Preview URL retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviewUrlResponse'
        '404':
          description: Container not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Container preview port not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # WebContainer-Compatible Endpoints
...
components:
  schemas:
    ContainerInfo:
      type: object
      required:
        - id
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique container identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        status:
          type: string
          enum: [creating, running, stopped, error]
          description: Current container status
          example: "running"
        port:
          type: integer
          description: WebSocket port for container communication
          example: 8342
        previewPort:
          type: integer
          description: Port for web application preview
          example: 3000
        websocketUrl:
          type: string
          format: uri
          description: WebSocket URL for real-time communication
          example: "ws://localhost:3001/container"
        previewUrl:
          type: string
          format: uri
          description: Preview URL for web applications
          example: "http://localhost:3000"
        createdAt:
          type: string
          format: date-time
          description: Container creation timestamp
          example: "2025-09-18T10:30:00Z"
    FileSystemTree:
      type: object
      description: Nested file system structure
      additionalProperties:
        $ref: '#/components/schemas/FileNode'
      example:
        "package.json":
          file:
            contents: |
              {
                "name": "my-app",
                "version": "1.0.0"
              }
        "src":
          directory:
            "index.js":
              file:
                contents: "console.log('Hello World');"
    FileNode:
      type: object
      description: File system node (file or directory)
      properties:
        file:
          type: object
          properties:
            contents:
              type: string
              description: File contents
        directory:
          $ref: '#/components/schemas/FileSystemTree'
      oneOf:
        - required: [file]
        - required: [directory]
...
    PreviewUrlResponse:
      type: object
      required:
        - url
        - containerId
        - port
      properties:
        url:
          type: string
          format: uri
          description: Preview URL
          example: "http://localhost:3000"
        containerId:
          type: string
          format: uuid
          description: Container ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        port:
          type: integer
          description: Preview port
          example: 3000
...
    ErrorResponse:
      type: object
      required:
        - message
        - statusCode
      properties:
        message:
          type: string
          description: Error message
          example: "Container not found"
        statusCode:
          type: integer
          description: HTTP status code
          example: 404
        error:
          type: string
          description: Error type
          example: "Not Found"
  securitySchemes: {}
...